event file rearrangement
- moved patrol directory into events_module directory
- moved short event class into own file and short directory within events_module directory
