+++ scripts/events_module/patrol/patrol_outcome.py
@@ -11,7 +11,7 @@ import pygame
 from scripts.events_module.handle_short_events import INJURY_GROUPS
 
 if TYPE_CHECKING:
-    from scripts.patrol.patrol import Patrol
+    from scripts.events_module.patrol.patrol import Patrol
 
 from scripts.cat.history import History
 from scripts.clan import HERBS
